+++ ceammc/CHANGELOG.md
@@ -12,6 +12,7 @@
   - list.map: list value mapping object added
   - flt.a-weight: A-weight amplitude/decibel value calculator
   - an.zero~: zero crossing count/rate/freq
+  - proto.whammy: Digitech Whammy pedal control
 - new properties:
   - @clip property added to system.cursor to clip cursor XY coords
   - @norm property added to ui.env to do envelope value normalization on input